app/models/image_analyzer.py
import cv2
import numpy as np
from PIL import Image
import requests
from io import BytesIO
import torch
from torchvision import transforms
from torchvision.models import resnet50, ResNet50_Weights

class ImageAnalyzer:
    def __init__(self):
        # Initialize the model for image classification
        self.model = resnet50(weights=ResNet50_Weights.DEFAULT)
        self.model.eval()
        if torch.cuda.is_available():
            self.model = self.model.cuda()
        
        # Image preprocessing
        self.transform = transforms.Compose([
            transforms.Resize(256),
            transforms.CenterCrop(224),
            transforms.ToTensor(),
            transforms.Normalize(
                mean=[0.485, 0.456, 0.406],
                std=[0.229, 0.224, 0.225]
            )
        ])

    def analyze(self, image_url):
        """
        Analyze image for potential manipulation or suspicious content
        Returns a score between 0 (likely authentic) and 1 (likely manipulated)
        """
        try:
            # Download and process image
            response = requests.get(image_url)
            img = Image.open(BytesIO(response.content)).convert('RGB')
            
            # Basic image analysis
            manipulation_score = self._detect_manipulation(img)
            content_score = self._analyze_content(img)
            
            # Combine scores (weighted average)
            final_score = 0.6 * manipulation_score + 0.4 * content_score
            
            return final_score
            
        except Exception as e:
            print(f"Error analyzing image: {str(e)}")
            return 0.5  # Neutral score on error

    def _detect_manipulation(self, img):
        """
        Detect potential image manipulation using various techniques
        """
        # Convert to OpenCV format
        img_cv = cv2.cvtColor(np.array(img), cv2.COLOR_RGB2BGR)
        
        # Basic manipulation detection
        # 1. Check for compression artifacts
        compression_score = self._check_compression_artifacts(img_cv)
        
        # 2. Check for edge inconsistencies
        edge_score = self._check_edge_inconsistencies(img_cv)
        
        # 3. Check for noise patterns
        noise_score = self._check_noise_patterns(img_cv)
        
        # Combine scores
        return (compression_score + edge_score + noise_score) / 3

    def _analyze_content(self, img):
        """
        Analyze image content for suspicious patterns
        """
        # Preprocess image for model
        img_tensor = self.transform(img).unsqueeze(0)
        if torch.cuda.is_available():
            img_tensor = img_tensor.cuda()
        
        # Get model predictions
        with torch.no_grad():
            outputs = self.model(img_tensor)
        
        # TODO: Implement content analysis logic
        # - Check for known fake image patterns
        # - Analyze image composition
        # - Look for suspicious elements
        
        return 0.5  # Placeholder return

    def _check_compression_artifacts(self, img):
        """
        Check for JPEG compression artifacts
        """
        # TODO: Implement compression artifact detection
        return 0.5

    def _check_edge_inconsistencies(self, img):
        """
        Check for edge inconsistencies that might indicate manipulation
        """
        # TODO: Implement edge analysis
        return 0.5

    def _check_noise_patterns(self, img):
        """
        Check for unusual noise patterns
        """
        # TODO: Implement noise pattern analysis
        return 0.5 