requirements.txt
flask==2.0.1
numpy==1.21.0
pandas==1.3.0
scikit-learn==0.24.2
torch==1.9.0
torchvision==0.10.0
transformers==4.11.3
pillow==8.3.1
tensorflow==2.6.0
opencv-python==4.5.3.56
python-dotenv==0.19.0
requests==2.26.0
flask-cors==3.0.10 